requirements.txt
streamlit>=1.24.0
torch>=2.0.0
transformers>=4.30.0
python-dotenv>=0.19.0